+++ src/upstream/types/platform.rs
@@ -1,0 +1,69 @@
+use serde::{Serialize, Deserialize};
+use strum_macros::{EnumString, EnumIter, Display};
+
+/// All identity platform.
+/// TODO: move this definition into `graph/vertex/identity`, since it is not specific to upstream.
+#[derive(
+    Serialize,
+    Deserialize,
+    Debug,
+    EnumString,
+    Clone,
+    Copy,
+    Display,
+    PartialEq,
+    Eq,
+    EnumIter,
+    Default,
+    async_graphql::Enum,
+)]
+pub enum Platform {
+    /// Twitter
+    #[strum(serialize = "twitter")]
+    #[serde(rename = "twitter")]
+    #[graphql(name = "twitter")]
+    Twitter,
+
+    /// Ethereum wallet `0x[a-f0-9]{40}`
+    #[strum(serialize = "ethereum", serialize = "eth")]
+    #[serde(rename = "ethereum")]
+    #[graphql(name = "ethereum")]
+    Ethereum,
+
+    /// NextID
+    #[strum(serialize = "nextid")]
+    #[serde(rename = "nextid")]
+    #[graphql(name = "nextid")]
+    NextID,
+
+    /// Keybase
+    #[strum(serialize = "keybase")]
+    #[serde(rename = "keybase")]
+    #[graphql(name = "keybase")]
+    Keybase,
+
+    /// Github
+    #[strum(serialize = "github")]
+    #[serde(rename = "github")]
+    #[graphql(name = "github")]
+    Github,
+
+    /// Reddit
+    #[strum(serialize = "reddit")]
+    #[serde(rename = "reddit")]
+    #[graphql(name = "reddit")]
+    Reddit,
+
+    /// Lens
+    #[strum(serialize = "lens")]
+    #[serde(rename = "lens")]
+    #[graphql(name = "lens")]
+    Lens,
+
+    /// Unknown
+    #[strum(serialize = "unknown")]
+    #[serde(rename = "unknown")]
+    #[graphql(name = "unknown")]
+    #[default]
+    Unknown,
+}
